requirements.txt
distro==1.8.0
et-xmlfile==1.1.0
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
tabula-py==2.6.0
